Sacando referencias desde otros lugares en elecciones para votainteligentemessages y votainteligenteanswers
